+++ Vaigachev/2/battlefield.h
@@ -1,0 +1,22 @@
+#pragma once
+#include "list.h"
+#include "Object.h"
+#include <memory>
+//#include "list.cpp" // ?-----?
+
+class battlefield
+{
+	int x_size;
+	int y_size;
+	List <Object*> *team1 = new List<Object*>;
+	List <Object*> *team2 = new List<Object*>;
+public:
+	battlefield();
+    ~battlefield();
+	void new_team(List<Object*> *new_team, ifstream& fin,const shared_ptr<int>);
+	void print();
+	int hit(_2dim, int);
+	Object* check_position(_2dim);
+	void del_from_position(_2dim);
+};
+
