--- Вайгачев/1/data.txt
@@ -1,10 +1,0 @@
-10 10
-4
-0 0 80
-1 1 100
-4 4 100
-5 5 100
-3
-8 8 100
-5 8 100
-8 5 99
