--- Вайгачев/2/data.txt
@@ -1,10 +1,0 @@
-10 10
-4
-2 2 80
-1 1 100
-4 4 100
-5 5 1
-3
-8 8 100
-5 8 100
-8 5 99
